Add support for second generation of Xiaomi climate sensors (#116)
This includes clocks and dedicated temperature/humidity sensors family
known as LYWSD02.

I got my hand on one of those literally for few hours and managed to
make it work with bt-mqtt-gateway pretty easily. There's much more to
implement, especially support for setting time, units (C/F) and maybe
even history. That said - with this code it's possible to read it but
it needs to be configured separately.

I structured code similarty to IBBQ one, as it's pretty similar use
case when we connect with bluepy.