fix unknown device name in command message
Device names are retrieved by the configuration file. If an MQTT command is
handled but contains an unknown device name, which is not in the configuration
file an KeyError was thrown and the application was terminated. This commit
checks if the device name is known, if not it logs an error message but will not
fail.